refactor: create protocols.reader module
